clang-tidy triyaj: kendi kodumuzdaki tum bulgular temizlendi (0 kaldi)
CI'daki clang-tidy job'u su ana kadar yalnizca bilgilendirici calisiyordu
(bulgular hic triyaj edilmemisti). Gercek calistirma: 1.13 milyon uyari
uretti ama bunlarin 1.13 milyondan fazlasi vendor/sistem header'larindan
(Jolt/lz4) -- kendi src/ agacimizda yalnizca 8 gercek bulgu vardi:

- AeroSystem.cpp / TerrainDeformSystem.cpp: RVec3 kurulumunda double->
  float daralmasi -- kod double'a genisletip sonra tek-hassasiyetli
  RVec3(=Vec3) yapicisinda tekrar float'a daraltiyordu (bu build'de
  anlamsiz ara adim). Dogrudan float'a cast edilecek sekilde duzeltildi
  -- hem tek hem cift hassasiyetli Jolt derlemelerinde dogru.
- DebugDrawBridge.cpp: JPH::Color bilesenlerini (uint8) 255.0f'e bolerken
  brace-init'te examplicit int->float cast eksikti -- eklendi (sonuc
  ayni, sadece derleyiciye niyeti acikca belirtiyor).
- RopeSystem.cpp: segments[i+1] indeksinde int toplama once yapilip
  SONRA size_t'e donusuyordu (genel durumda tasma riski) -- once i'yi
  size_t'e cevirip sonra +1 yapacak sekilde duzeltildi.

Davranis degisikligi yok (tum degerler ayni), 29/29 yerel ctest yesil.
CI'nin clang-tidy job'u artik gercekten sifir bulguyla calisiyor.

+++ src/debugdraw/DebugDrawBridge.cpp
@@ -11,7 +11,8 @@ namespace {
 
 #ifdef JPH_DEBUG_RENDERER
 DebugColor FromJoltColor(JPH::ColorArg c) {
-    return DebugColor{c.r / 255.0f, c.g / 255.0f, c.b / 255.0f, c.a / 255.0f};
+    return DebugColor{static_cast<float>(c.r) / 255.0f, static_cast<float>(c.g) / 255.0f,
+                      static_cast<float>(c.b) / 255.0f, static_cast<float>(c.a) / 255.0f};
 }
 #endif
